lint: config ktlint and rename files
