+++ presentation/src/main/java/dev/jamile/presentation/navigation/Navigation.kt
@@ -5,7 +5,7 @@ import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
 import dev.jamile.presentation.ui.favorites.FavoritesScreen
-import dev.jamile.presentation.ui.game_detail.GameDetailScreen
+import dev.jamile.presentation.ui.gamedetail.GameDetailScreen
 import dev.jamile.presentation.ui.home.HomeScreen
 import dev.jamile.presentation.ui.search.SearchScreen
