+++ presentation/src/main/java/dev/jamile/presentation/ui/gamedetail/GameDetailHeader.kt
@@ -1,4 +1,4 @@
-package dev.jamile.presentation.ui.game_detail
+package dev.jamile.presentation.ui.gamedetail
 
 import androidx.compose.foundation.ScrollState
 import androidx.compose.foundation.layout.Box
